Feat: 데이터 준비중 메세지
